# openapi/v0.yaml
openapi: 3.1.0
info:
  title: Trix API
  version: 0.1.0
  summary: Scaffolded v0 API for the Trix backend
servers:
  - url: http://localhost:8080
tags:
  - name: system
  - name: auth
  - name: accounts
  - name: devices
  - name: chats
  - name: blobs
paths:
  /v0/system/health:
    get:
      operationId: getSystemHealth
      tags: [system]
      responses:
        '200':
          description: Service health
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /v0/system/version:
    get:
      operationId: getSystemVersion
      tags: [system]
      responses:
        '200':
          description: Build information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'
  /v0/auth/challenge:
    post:
      operationId: createAuthChallenge
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthChallengeRequest'
      responses:
        '200':
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthChallengeResponse'
  /v0/auth/session:
    post:
      operationId: createAuthSession
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthSessionRequest'
      responses:
        '200':
          description: Session token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
  /v0/accounts:
    post:
      operationId: createAccount
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '200':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAccountResponse'
  /v0/accounts/me:
    get:
      operationId: getCurrentAccount
      tags: [accounts]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated account profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProfileResponse'
  /v0/devices:
    get:
      operationId: listDevices
      tags: [devices]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Devices for the authenticated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
  /v0/chats:
    get:
      operationId: listChats
      tags: [chats]
      responses:
        '501':
          $ref: '#/components/responses/NotImplemented'
    post:
      operationId: createChat
      tags: [chats]
      responses:
        '501':
          $ref: '#/components/responses/NotImplemented'
  /v0/blobs/uploads:
    post:
      operationId: createBlobUpload
      tags: [blobs]
      responses:
        '501':
          $ref: '#/components/responses/NotImplemented'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotImplemented:
      description: Endpoint scaffold exists but behavior is not implemented yet.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ServiceStatus:
      type: string
      enum: [ok, degraded]
    HealthResponse:
      type: object
      required: [service, status, version, uptime_ms]
      properties:
        service:
          type: string
        status:
          $ref: '#/components/schemas/ServiceStatus'
        version:
          type: string
        uptime_ms:
          type: integer
          format: int64
    VersionResponse:
      type: object
      required: [service, version]
      properties:
        service:
          type: string
        version:
          type: string
        git_sha:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    DeviceStatus:
      type: string
      enum: [pending, active, revoked]
    CreateAccountRequest:
      type: object
      required:
        - profile_name
        - device_display_name
        - platform
        - credential_identity_b64
        - account_root_pubkey_b64
        - account_root_signature_b64
        - transport_pubkey_b64
      properties:
        handle:
          type: string
          nullable: true
        profile_name:
          type: string
        profile_bio:
          type: string
          nullable: true
        device_display_name:
          type: string
        platform:
          type: string
        credential_identity_b64:
          type: string
        account_root_pubkey_b64:
          type: string
        account_root_signature_b64:
          type: string
        transport_pubkey_b64:
          type: string
    CreateAccountResponse:
      type: object
      required: [account_id, device_id, account_sync_chat_id]
      properties:
        account_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        account_sync_chat_id:
          type: string
          format: uuid
    AuthChallengeRequest:
      type: object
      required: [device_id]
      properties:
        device_id:
          type: string
          format: uuid
    AuthChallengeResponse:
      type: object
      required: [challenge_id, challenge_b64, expires_at_unix]
      properties:
        challenge_id:
          type: string
          format: uuid
        challenge_b64:
          type: string
        expires_at_unix:
          type: integer
          format: int64
    AuthSessionRequest:
      type: object
      required: [device_id, challenge_id, signature_b64]
      properties:
        device_id:
          type: string
          format: uuid
        challenge_id:
          type: string
          format: uuid
        signature_b64:
          type: string
    AuthSessionResponse:
      type: object
      required: [access_token, expires_at_unix, account_id, device_status]
      properties:
        access_token:
          type: string
        expires_at_unix:
          type: integer
          format: int64
        account_id:
          type: string
          format: uuid
        device_status:
          $ref: '#/components/schemas/DeviceStatus'
    AccountProfileResponse:
      type: object
      required: [account_id, profile_name, device_id, device_status]
      properties:
        account_id:
          type: string
          format: uuid
        handle:
          type: string
          nullable: true
        profile_name:
          type: string
        profile_bio:
          type: string
          nullable: true
        device_id:
          type: string
          format: uuid
        device_status:
          $ref: '#/components/schemas/DeviceStatus'
    DeviceSummary:
      type: object
      required: [device_id, display_name, platform, device_status]
      properties:
        device_id:
          type: string
          format: uuid
        display_name:
          type: string
        platform:
          type: string
        device_status:
          $ref: '#/components/schemas/DeviceStatus'
    DeviceListResponse:
      type: object
      required: [account_id, devices]
      properties:
        account_id:
          type: string
          format: uuid
        devices:
          type: array
          items:
            $ref: '#/components/schemas/DeviceSummary'